Separate complex from simple dependencies and create a test
