+++ src/renderer/src/stories/pages/guided-mode/GuidedHome.js
@@ -3,7 +3,7 @@ import { Page } from "../Page.js";
 import { ProgressCard } from "./ProgressCard.js";
 
 import { startLottie } from "../../../dependencies/globals.js";
-import * as progress from "../../../progress.js";
+import * as progress from "../../../progress/index.js";
 import { newDataset } from "../../../../assets/lotties/index.js";
 
 export class GuidedHomePage extends Page {
